+++ include/ggml/tpool.h
@@ -2,7 +2,7 @@
 
 #if defined(_WIN32)
 #include <windows.h>
-#include "pthreads.h"
+#include "pthreads-win32.h"
 #else
 #include <pthreads.h>
 #endif
